// packages/http/types/hooks.d.ts
type HookHandlerDoneFunction = <TError extends Error = XufaError>(err?: TError) => void;

interface RequestPayload extends Readable {
  receivedEncodedLength?: number;
}
// ...
export interface preHandlerAsyncHookHandler<
  RawServer extends RawServerBase = RawServerDefault,
  RawRequest extends RawRequestDefaultExpression<RawServer> = RawRequestDefaultExpression<RawServer>,
  RawReply extends RawReplyDefaultExpression<RawServer> = RawReplyDefaultExpression<RawServer>,
  RouteGeneric extends RouteGenericInterface = RouteGenericInterface,
  ContextConfig = ContextConfigDefault,
  SchemaCompiler extends XufaSchema = XufaSchema,
  TypeProvider extends XufaTypeProvider = XufaTypeProviderDefault,
  Logger extends XufaBaseLogger = XufaBaseLogger,
> {
  (
    this: XufaInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>,
    request: XufaRequest<RouteGeneric, RawServer, RawRequest, SchemaCompiler, TypeProvider, ContextConfig, Logger>,
    reply: XufaReply<RouteGeneric, RawServer, RawRequest, RawReply, ContextConfig, SchemaCompiler, TypeProvider>
  ): Promise<unknown>;
}
// ...
interface DoneFuncWithErrOrRes {
  (): void;
  <TError extends Error = XufaError>(err: TError): void;
  (err: null, res: unknown): void;
}